Began refactoring, fixed seg fault
Moved the math subroutines into their own module. Ultimately this is a semantic change, it doesn't impact execution, but it does establish some *ahem* modularity. In addition, I fixed a segmentation fault with the factorial command. I'm unsure how this did not affect execution when completing the assignment, but has been fixed now. It was a result of listTwo being check uninitialized.